Crate pallet_investments

Re-exports

pub use weights::WeightInfo;

pub use pallet::*;

Modules

pallet

The module that hosts all the FRAME types needed to add this pallet to a runtime.

weights

Enums

CollectOutcome

A signaler, showing if the collect call actually collected all Closed orders and the investor can create a new investment or if the investor has to call collect() again.

CollectType

Defining how the collect logic runs. CollectType::Closing will ensure, that all unfulfilled investments are returned to the user account. CollectType::Overflowing will ensure, that all unfilfilled investments are moved into the next active order for this investment.

OrderType

The enum we parse to PreConditions so the runtime can make an educated decision about this investment

Type Aliases

OrderOf

A newtype for Order