Trait pallet_investments::pallet::Config

pub trait Config: Config where
    <Self::Accountant as InvestmentAccountant<Self::AccountId>>::InvestmentInfo: InvestmentProperties<Self::AccountId, Currency = <<Self as Config>::Tokens as Inspect<<Self as Config>::AccountId>>::AssetId>, {
    type Event: From<Event<Self>> + IsType<<Self as Config>::Event>;
    type InvestmentId: Member + Parameter + Copy;
    type Accountant: InvestmentAccountant<Self::AccountId, Error = DispatchError, InvestmentId = Self::InvestmentId, Amount = Self::Amount>;
    type Amount: Member + Parameter + AtLeast32BitUnsigned + Default + Copy + MaxEncodedLen + FixedPointOperand + From<u64> + From<u128> + TryInto<u64>;
    type BalanceRatio: Member + Parameter + Default + Copy + FixedPointNumber<Inner = Self::Amount>;
    type MaxOutstandingCollects: Get<u64>;
    type Tokens: Mutate<Self::AccountId> + Inspect<Self::AccountId, Balance = Self::Amount> + Transfer<Self::AccountId>;
    type PreConditions: PreConditions<OrderType<Self::AccountId, Self::InvestmentId, Self::Amount>, Result = bool>;
    type WeightInfo: WeightInfo;
}

Configure the pallet by specifying the parameters and types on which it depends.

Required Associated Types

type Event: From<Event<Self>> + IsType<<Self as Config>::Event>

Because this pallet emits events, it depends on the runtime’s definition of an event.

type InvestmentId: Member + Parameter + Copy

The underlying investments one can invest into

type Accountant: InvestmentAccountant<Self::AccountId, Error = DispatchError, InvestmentId = Self::InvestmentId, Amount = Self::Amount>

Something that knows how to handle accounting for the given investments and provides metadata about them

type Amount: Member + Parameter + AtLeast32BitUnsigned + Default + Copy + MaxEncodedLen + FixedPointOperand + From<u64> + From<u128> + TryInto<u64>

A representation for an investment or redemption. Usually this is equal to the known Balance type of a system.

type BalanceRatio: Member + Parameter + Default + Copy + FixedPointNumber<Inner = Self::Amount>

A fixed-point number which represents the value of one currency type in terms of another.

type MaxOutstandingCollects: Get<u64>

The bound on how many fulfilled orders we cache until the user needs to collect them.

type Tokens: Mutate<Self::AccountId> + Inspect<Self::AccountId, Balance = Self::Amount> + Transfer<Self::AccountId>

Something that can handle payments and transfers of currencies

type PreConditions: PreConditions<OrderType<Self::AccountId, Self::InvestmentId, Self::Amount>, Result = bool>

A possible check if investors fulfill every condition to invest into a given investment

type WeightInfo: WeightInfo

The weight information for this pallet extrinsics.