Type Alias pallet_investments::pallet::Module

pub type Module<T> = Pallet<T>;

👎Deprecated: use Pallet instead

Type alias to Pallet, to be used by construct_runtime.

Generated by pallet attribute macro.

Aliased Type

struct Module<T>(/* private fields */);

Implementations

impl<T: Config> Pallet<T>

pub fn update_invest_order( origin: OriginFor<T>, investment_id: T::InvestmentId, amount: T::Amount ) -> DispatchResult

Update an order to invest into a given investment.

If the requested amount is greater than the current investment order, the balance will be transferred from the calling account to the pool. If the requested amount is less than the current order, the balance will be transferred from the pool to the calling account.

pub fn update_redeem_order( origin: OriginFor<T>, investment_id: T::InvestmentId, amount: T::Amount ) -> DispatchResult

Update an order to redeem from a given investment.

If the requested amount is greater than the current investment order, the balance will be transferred from the calling account to the pool. If the requested amount is less than the current order, the balance will be transferred from the pool to the calling account.

pub fn collect_investments( origin: OriginFor<T>, investment_id: T::InvestmentId ) -> DispatchResultWithPostInfo

Collect the results of a user’s invest orders for the given investment. If any amounts are not fulfilled they are directly appended to the next active order for this investment.

pub fn collect_redemptions( origin: OriginFor<T>, investment_id: T::InvestmentId ) -> DispatchResultWithPostInfo

Collect the results of a user’s redeem orders for the given investment. If any amounts are not fulfilled they are directly appended to the next active order for this investment.

pub fn collect_investments_for( origin: OriginFor<T>, who: T::AccountId, investment_id: T::InvestmentId ) -> DispatchResultWithPostInfo

Collect the results of another users invest orders for the given investment. If any amounts are not fulfilled they are directly appended to the next active order for this investment.

pub fn collect_redemptions_for( origin: OriginFor<T>, who: T::AccountId, investment_id: T::InvestmentId ) -> DispatchResultWithPostInfo

Collect the results of another users redeem orders for the given investment. If any amounts are not fulfilled they are directly appended to the next active order for this investment.

impl<T: Config> Pallet<T>

pub fn invest_order_id<KArg>(k: KArg) -> OrderIdwhere KArg: EncodeLike<T::InvestmentId>,

impl<T: Config> Pallet<T>

pub fn redeem_order_id<KArg>(k: KArg) -> OrderIdwhere KArg: EncodeLike<T::InvestmentId>,

impl<T: Config> Pallet<T>

pub fn acc_active_invest_order<KArg>(k: KArg) -> TotalOrder<T::Amount>where KArg: EncodeLike<T::InvestmentId>,

impl<T: Config> Pallet<T>

pub fn acc_active_redeem_order<KArg>(k: KArg) -> TotalOrder<T::Amount>where KArg: EncodeLike<T::InvestmentId>,

impl<T: Config> Pallet<T>

pub fn acc_in_processing_invest_order<KArg>( k: KArg ) -> Option<TotalOrder<T::Amount>>where KArg: EncodeLike<T::InvestmentId>,

impl<T: Config> Pallet<T>

pub fn acc_in_processing_redeem_order<KArg>( k: KArg ) -> Option<TotalOrder<T::Amount>>where KArg: EncodeLike<T::InvestmentId>,

impl<T: Config> Pallet<T>

pub fn cleared_invest_order<KArg1, KArg2>( k1: KArg1, k2: KArg2 ) -> Option<FulfillmentWithPrice<T::BalanceRatio>>where KArg1: EncodeLike<T::InvestmentId>, KArg2: EncodeLike<OrderId>,

impl<T: Config> Pallet<T>

pub fn cleared_redeem_order<KArg1, KArg2>( k1: KArg1, k2: KArg2 ) -> Option<FulfillmentWithPrice<T::BalanceRatio>>where KArg1: EncodeLike<T::InvestmentId>, KArg2: EncodeLike<OrderId>,

impl<T: Config> Pallet<T>

pub fn acc_payout_invest( collection: &mut InvestCollection<T::Amount>, fulfillment: &FulfillmentWithPrice<T::BalanceRatio> ) -> Result<T::Amount, DispatchError>

Increments an accounts’ investment payout amount based on the remaining amount and the fulfillment price.

Returns the amount by which was incremented.

pub fn acc_payout_redeem( collection: &mut RedeemCollection<T::Amount>, fulfillment: &FulfillmentWithPrice<T::BalanceRatio> ) -> Result<T::Amount, DispatchError>

Increments an accounts’ redemption payout amount based on the remaining amount and the fulfillment price.

Returns the amount by which was incremented.

pub fn acc_remaining_redeem( collection: &mut RedeemCollection<T::Amount>, fulfillment: &FulfillmentWithPrice<T::BalanceRatio> ) -> DispatchResult

Decrements an accounts’ remaining redemption amount based on the fulfillment price.

pub fn acc_remaining_invest( collection: &mut InvestCollection<T::Amount>, fulfillment: &FulfillmentWithPrice<T::BalanceRatio> ) -> DispatchResult

Decrements an accounts’ remaining investment amount based on the fulfillment price.